3D_Icon
This project includes three-dimensional icon designs created using HTML and CSS. I aimed to create modern and dynamic icons using web technologies. Using the power of CSS, I added depth to the icon and achieved a realistic 3D effect.

By playing with CSS features such as transform, perspective and box-shadow, I designed icons that are responsive and impressive to user interaction. Additionally, the icon's ability to cast different shades and colors depending on the light source allowed it to imitate objects in the real world.

Thanks to this project, I discovered how to develop 3D icons that provide a rich experience in terms of both aesthetics and functionality in user interfaces. I also gained a better understanding of the importance of optimizing browser performance and maintaining minimalism in design.

The main purpose of the project was to show what can be done with HTML and CSS while designing simple but eye-catching icons and at the same time improve my own design skills.